+++ mozilla_django_oidc/middleware.py
@@ -11,7 +11,7 @@ from django.utils.module_loading import import_string
 
 from mozilla_django_oidc.auth import OIDCAuthenticationBackend
 from mozilla_django_oidc.utils import (absolutify,
-                                       add_state_and_nonce_to_session,
+                                       add_state_and_verifier_and_nonce_to_session,
                                        import_from_settings)
 
 from urllib.parse import quote, urlencode
@@ -156,7 +156,7 @@ class SessionRefresh(MiddlewareMixin):
                 'nonce': nonce
             })
 
-        add_state_and_nonce_to_session(request, state, params)
+        add_state_and_verifier_and_nonce_to_session(request, state, params)
 
         request.session['oidc_login_next'] = request.get_full_path()