+++ mozilla_django_oidc/utils.py
@@ -2,8 +2,13 @@ import logging
 import time
 import warnings
 
+# Make it obvious that these aren't the usual base64 functions 
+import josepy.b64
+
 from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
+
+from hashlib import sha256
 
 from urllib.request import parse_http_list, parse_keqv_list
 
@@ -54,15 +59,70 @@ def is_authenticated(user):
     return user.is_authenticated
 
 
-def add_state_and_nonce_to_session(request, state, params):
+def base64_url_encode(bytes_like_obj) -> str:
+    """Return a URL-Safe, base64 encoded version of bytes_like_obj
+
+    Implements base64urlencode as described in https://datatracker.ietf.org/doc/html/rfc7636#appendix-A
     """
-    Stores the `state` and `nonce` parameters in a session dictionary including the time when it
-    was added. The dictionary can contain multiple state/nonce combinations to allow parallel
-    logins with multiple browser sessions.
-    To keep the session space to a reasonable size, the dictionary is kept at 50 state/nonce
-    combinations maximum.
+
+    s = josepy.b64.b64encode(bytes_like_obj) # josepy base64 encoder (strips '='s padding)
+    s = s.replace('+', '-') # 62nd char of encoding
+    s = s.replace('/', '_') # 63rd char of encoding
+
+    return s
+
+
+def base64_url_decode(string_like_obj) -> bytes:
+    """Return the bytes encoded in a URL-Safe, base64 encoded string
+    Implements inverse of base64urlencode as described in https://datatracker.ietf.org/doc/html/rfc7636#appendix-A
+    This function is not used by the OpenID client; it's just for testing PKCE related functions.
+    """
+    s = string_like_obj
+
+    s = s.replace('_', '/') # 63rd char of encoding
+    s = s.replace('-', '+') # 62nd char of encoding
+    b = josepy.b64.b64decode(s) # josepy base64 encoder (decodes without '='s padding)
+
+    return b
+
+
+def generate_code_challenge(code_verifier, method) -> str:
+    """Return a code_challege, which proves knowledge of the code_verifier.
+    The code challenge is generated according method which must be one
+    of the methods defined in https://datatracker.ietf.org/doc/html/rfc7636#section-4.2:
+    - plain
+      code_challenge = code_verifier
+    - S256
+      code_challenge = BASE64URL-ENCODE(SHA256(ASCII(code_verifier)))
+    """
+
+    if method == 'plain':
+        return code_verifier
+
+    elif method == 'S256':
+        return base64_url_encode(sha256(code_verifier.encode('ascii')).digest())
+
+    else:
+        raise ValueError("code challenge method must be 'plain' or 'S256'.")
+
+
+def add_state_and_verifier_and_nonce_to_session(request, state, params, code_verifier=None):
+    """
+    Stores the `state` and `nonce` parameters and an optional `code_verifier` (for PKCE) in a session dictionary 
+    which maps `state` -> {nonce, code_verifier}.  Each entry included the time when it
+    was added. The dictionary can contain multiple state -> {nonce, code_verifier}
+    mappings to allow parallel logins with multiple browser sessions.
+    To keep the session space to a reasonable size, the dictionary is kept at 50 state -> {nonce, code_verifier}
+    mappings maximum.
     """
     nonce = params.get('nonce')
+
+    # OPs supporting PKCE will require `code_verifier` to be sent to the token
+    # endpoint if `code_challenge` is sent to the authentication endpoint.
+    # Make sure that `code_challenge` and `code_verifier` are both specified
+    # or neither is.
+    assert ('code_challenge' in params) == (code_verifier is not None)
+    
 
     # Store Nonce with the State parameter in the session "oidc_states" dictionary.
     # The dictionary can store multiple State/Nonce combinations to allow parallel
@@ -94,6 +154,7 @@ def add_state_and_nonce_to_session(request, state, params):
             del request.session['oidc_states'][oldest_state]
 
     request.session['oidc_states'][state] = {
+        'code_verifier': code_verifier,
         'nonce': nonce,
         'added_on': time.time(),
     }